requirements.txt
pyTelegramBotAPI
pandas
python-dotenv
